% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_convert_func_syntax.R
\name{.cfs_find_matching_paren}
\alias{.cfs_find_matching_paren}
\title{Internal: find the matching closing parenthesis}
\usage{
.cfs_find_matching_paren(chars, open_pos)
}
\description{
Internal: find the matching closing parenthesis
}
\keyword{internal}
