% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_convert_func_syntax.R
\name{.cfs_transform_text}
\alias{.cfs_transform_text}
\title{Internal: iterative transform (handles nesting via multiple passes)}
\usage{
.cfs_transform_text(text, direction)
}
\description{
Internal: iterative transform (handles nesting via multiple passes)
}
\keyword{internal}
