% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_convert_int_literals.R
\name{.cil_process_text}
\alias{.cil_process_text}
\title{Internal: character-level integer-literal rewrite}
\usage{
.cil_process_text(text)
}
\description{
Internal: character-level integer-literal rewrite
}
\keyword{internal}
