% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\name{.mfae_is_dots}
\alias{.mfae_is_dots}
\title{Is this node a literal \code{...}?}
\usage{
.mfae_is_dots(pd, row)
}
\description{
\code{row} is the row of the expression node to inspect.
}
\keyword{internal}
