% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\docType{data}
\name{.mfae_skipped_ops}
\alias{.mfae_skipped_ops}
\title{Operators and special forms that are never transformed}
\format{
An object of class \code{character} of length 35.
}
\usage{
.mfae_skipped_ops
}
\description{
Operators and special forms that are never transformed
}
\keyword{internal}
