% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\name{.mfae_walk}
\alias{.mfae_walk}
\title{Main recursive walker: dispatch per expression type}
\usage{
.mfae_walk(expr, skip_fns = NULL)
}
\description{
Main recursive walker: dispatch per expression type
}
\keyword{internal}
