% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\name{.mfae_walk_brace}
\alias{.mfae_walk_brace}
\title{Walk a \code{\{} block expression}
\usage{
.mfae_walk_brace(expr, skip_fns = NULL)
}
\description{
Walk a \code{\{} block expression
}
\keyword{internal}
