% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\name{.mfae_walk_children}
\alias{.mfae_walk_children}
\title{Recursively walk children without changing the call head}
\usage{
.mfae_walk_children(expr, skip_fns = NULL)
}
\description{
Recursively walk children without changing the call head
}
\keyword{internal}
