% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\name{.mfae_walk_expr}
\alias{.mfae_walk_expr}
\title{Walk an expression object (vector of top-level expressions)}
\usage{
.mfae_walk_expr(expr, skip_fns = NULL)
}
\description{
Walk an expression object (vector of top-level expressions)
}
\keyword{internal}
