% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\name{.mfae_walk_function_def}
\alias{.mfae_walk_function_def}
\title{Walk a \code{function} definition expression}
\usage{
.mfae_walk_function_def(expr, skip_fns = NULL)
}
\description{
Walk a \code{function} definition expression
}
\keyword{internal}
