% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\name{.mfae_walk_paren}
\alias{.mfae_walk_paren}
\title{Walk a parenthesised expression}
\usage{
.mfae_walk_paren(expr, skip_fns = NULL)
}
\description{
Walk a parenthesised expression
}
\keyword{internal}
