% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\name{.mfae_walk_repeat}
\alias{.mfae_walk_repeat}
\title{Walk a \code{repeat} loop expression}
\usage{
.mfae_walk_repeat(expr, skip_fns = NULL)
}
\description{
Walk a \code{repeat} loop expression
}
\keyword{internal}
