% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_make_func_arg_explicit.R
\name{.mfae_walk_while}
\alias{.mfae_walk_while}
\title{Walk a \code{while} loop expression}
\usage{
.mfae_walk_while(expr, skip_fns = NULL)
}
\description{
Walk a \code{while} loop expression
}
\keyword{internal}
