% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_add_global_rbuildignore.R
\name{GLOBAL_RBUILDIGNORE_PATTERN}
\alias{GLOBAL_RBUILDIGNORE_PATTERN}
\title{Global \code{.Rbuildignore} patterns}
\usage{
GLOBAL_RBUILDIGNORE_PATTERN
}
\description{
Default regular-expression patterns used by \code{\link[=add_global_rbuildignore]{add_global_rbuildignore()}}.
}
