% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_open_cffinit.R
\name{open_cffinit}
\alias{open_cffinit}
\title{Open the CFF initializer}
\usage{
open_cffinit(...)
}
\arguments{
\item{...}{Unused arguments, for future extensibility.}
}
\value{
Invisible \code{TRUE}, called for side effects.
}
\description{
Opens the CFF (Citation File Format) initializer web application in the
default browser.
}
\seealso{
\code{\link[=browseURL]{browseURL()}}
}
\concept{open functions}
