% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_use_vendor.R
\name{repo_spec_parse}
\alias{repo_spec_parse}
\title{Helper: parse a GitHub repo spec from user input}
\usage{
repo_spec_parse(pkg, branch = "main")
}
\description{
Helper: parse a GitHub repo spec from user input
}
\keyword{internal}
