% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-strict-list.R
\name{strict_list}
\alias{strict_list}
\title{Create a named list with strict checks.}
\usage{
strict_list(..., .name = NULL)
}
\arguments{
\item{...}{the named values used to construct the list}

\item{.name}{an optional description for the list, used in error messages}
}
\description{
Accessing an unknown name (using any of \code{$}, \code{[} or \code{[[}) will throw an error
rather than return NULL.
}
