% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/46_use_r_v4.1.0.R
\name{use_r_v4.1.0}
\alias{use_r_v4.1.0}
\title{Add a minimum \code{R} version dependency to a package}
\usage{
use_r_v4.1.0(path = NULL, ...)
}
\arguments{
\item{path}{Path to the package root. If \code{NULL} (the default), the current
working directory is used.}

\item{...}{Must be empty. Reserved for future arguments.}
}
\value{
Invisibly returns \code{NULL}, called for side effects.
}
\description{
Adds \verb{R (>= 4.1.0)} to the \code{Depends} field of a package's \code{DESCRIPTION} file.
This sets a minimum R version requirement for your package. Then you can use
\verb{\\()} and \verb{|>} syntax in your package.
\itemize{
\item Requires that the target directory is an R package root (contains a
\code{DESCRIPTION} file).
\item Calls \code{\link[usethis:use_package]{usethis::use_package()}} to add the dependency.
}
}
\examples{
\donttest{
tmpdir <- tempdir()
usethis::create_package(path = tmpdir)
use_r_v4.1.0(path = tmpdir)
}
}
