% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/43_use_vendor.R
\name{vendor_declare_source_license}
\alias{vendor_declare_source_license}
\title{Helper: set up inst/vendor/pkg/ with LICENSE files and README}
\usage{
vendor_declare_source_license(
  path = NULL,
  owner = character(1L),
  repo = character(1L),
  branch = character(1L),
  repo_url = character(1L)
)
}
\description{
Helper: set up inst/vendor/pkg/ with LICENSE files and README
}
\keyword{internal}
