% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_use_vendor.R
\name{vendor_update_desc}
\alias{vendor_update_desc}
\title{Helper: update DESCRIPTION with vendor authors and copyright}
\usage{
vendor_update_desc(
  path = ".",
  author_info = character(1L),
  repo = character(1L),
  repo_url = character(1L)
)
}
\description{
Helper: update DESCRIPTION with vendor authors and copyright
}
\keyword{internal}
