% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_render_rmd.R
\name{render_rmd}
\alias{render_rmd}
\title{Render an R Markdown or R document to Markdown format}
\usage{
render_rmd(path = NULL, output_format = "md_document", ...)
}
\arguments{
\item{path}{Path to the input file. If NULL and rstudioapi is available,
uses the currently active document in RStudio.}

\item{output_format}{Output format to render to. Defaults to "md_document".}

\item{...}{Additional arguments passed to rmarkdown::render.}
}
\value{
The output file path from rmarkdown::render.
}
\description{
Render an R Markdown or R document to Markdown format
}
\examples{
\dontrun{
rlang::is_installed("rmarkdown")
tmp <- tempfile(fileext = ".Rmd")
writeLines(c("---", "title: Test", "---", "", "Hello, world!"), tmp)
render_rmd(tmp)
}
}
