% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_pkg_description.R
\docType{package}
\name{rpkgkit-package}
\alias{rpkgkit}
\alias{rpkgkit-package}
\title{rpkgkit: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
