% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkgkit-package.R
\docType{package}
\name{rpkgkit-package}
\alias{rpkgkit-package}
\alias{rpkgkit}
\title{Create and Maintain R Packages}
\description{
Utilities for R package development including NEWS.md
management, standalone file creation, and code formatting. Supports popular
development workflows and integrates with 'usethis' and 'RStudio'. Includes
helper functions for renaming functions and detecting common coding errors.
}
\section{License}{

MIT + file LICENSE
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/rpkgkit}
  \item Report bugs at \url{https://github.com/WangLabCSU/rpkgkit/issues}
}

}
\author{
\strong{Maintainer}: Yuxi Yang \email{15364051195@163.com}

Authors:
\itemize{
  \item Yuxi Yang \email{15364051195@163.com}
  \item Jacob Scott (Author of the included pedant code (https://github.com/wurli/pedant)) [copyright holder]
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (Author of the included pkgdev code (https://github.com/dieghernan/pkgdev)) [copyright holder]
}

Other contributors:
\itemize{
  \item Christopher T. Kenny (Contributor to the included pedant code) [contributor, copyright holder]
  \item Sebastian Lammers (Contributor to the included pedant code) [contributor, copyright holder]
}

}
\keyword{internal}
