% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_pkg_description.R
\docType{package}
\name{rpkgkit-package}
\alias{rpkgkit}
\alias{rpkgkit-package}
\title{A Package for R Package Development}
\description{
A Package for R Package Development
}
\author{
\strong{Maintainer}: Yuxi Yang \email{15364051195@163.com}

}
\keyword{internal}
