% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_pkg_description.R
\docType{package}
\name{rpkgkit-package}
\alias{rpkgkit}
\alias{rpkgkit-package}
\title{A Package for R Package Development}
\description{
A Package for R Package Development
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/rpkgkit}
  \item Report bugs at \url{https://github.com/WangLabCSU/rpkgkit/issues}
}

}
\keyword{internal}
