% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_pkg_description.R
\docType{package}
\name{rpkgkit-package}
\alias{rpkgkit}
\alias{rpkgkit-package}
\title{A Package for R Package Development}
\description{
A Package for R Package Development
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/rpkgkit}
  \item Report bugs at \url{https://github.com/WangLabCSU/rpkgkit/issues}
}

}
\author{
\strong{Maintainer}: Yuxi Yang \email{15364051195@163.com}

Authors:
\itemize{
  \item Jacob Scott (Author of the included pedant code (https://github.com/wurli/pedant)) [copyright holder]
}

Other contributors:
\itemize{
  \item Christopher T. Kenny (Contributor to the included pedant code) [contributor, copyright holder]
  \item Sebastian Lammers (Contributor to the included pedant code) [contributor, copyright holder]
}

}
\keyword{internal}
