% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_pkg_description.R
\docType{package}
\name{rpkgkit-package}
\alias{rpkgkit}
\alias{rpkgkit-package}
\title{rpkgkit: Provides A Set of Tools to Help You Create A Package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yuxi Yang \email{15364051195@163.com}

}
\keyword{internal}
