% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_use_vendor.R
\name{vendor_desc_authors}
\alias{vendor_desc_authors}
\title{Helper: extract author info from vendor DESCRIPTION}
\usage{
vendor_desc_authors(desc)
}
\description{
Helper: extract author info from vendor DESCRIPTION
}
\keyword{internal}
