% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_use_vendor.R
\name{vendor_desc_license}
\alias{vendor_desc_license}
\title{Helper: read license SPDX from the vendor's DESCRIPTION License field}
\usage{
vendor_desc_license(desc)
}
\description{
Helper: read license SPDX from the vendor's DESCRIPTION License field
}
\keyword{internal}
